requirements.txt
requests
python-dotenv
opencage
openai